% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{aghqInput}
\alias{aghqInput}
\title{Set aghq::quad parameters for model fitting.}
\usage{
aghqInput(...)
}
\arguments{
\item{...}{See details.}
}
\value{
A list of parameters for use by aghq::aghq.
}
\description{
Set aghq::quad parameters for model fitting.
}
\examples{
aghqInput()
aghqInput(k=5)
}
