% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{bs_effect}
\alias{bs_effect}
\title{Specify a polynomial fixed effect.}
\usage{
bs_effect(knots, degree = 3, ref_value = knots[[2]][1])
}
\arguments{
\item{poly_degree}{Degree of the polynomial.}
}
\value{
A list specifying a polynomial fixed effect.
}
\description{
Specify a polynomial fixed effect.
}
