% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ccDesign}
\alias{ccDesign}
\title{Specify the case-crossover design to be used by fitModel.}
\usage{
ccDesign(...)
}
\arguments{
\item{...}{See details.}
}
\description{
Function for specifying the case control design.
}
\details{
A default design is returned by \code{ccDesign()}. This latter can be changed, e.g., with \code{ccDesign(lag = 2)}.
Arguments of interest are \code{scheme} ("unidirectional", "bidirectional", "time stratified"),
\code{n_control} (number of control days, for uni- and bidirectional schemes), \code{lag} (e.g., 7 for ensuring that
case and control fall on the same weekday) and \code{stratum_rule} (for time stratified schemes, the only rule
implemented is "month").
}
\examples{
ccDesign()
}
