% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ccModel}
\alias{ccModel}
\title{Constructor for ccModel (S3) objects.}
\usage{
ccModel(
  response,
  time_index,
  fixed,
  random,
  overdispersion,
  design,
  aghq_input
)
}
\arguments{
\item{response}{Name of the response variable (e.g., count).}

\item{time_index}{Name of the time variable (e.g., date or time)}

\item{fixed}{Named list of fixed effects. Each name should correspond to a variable in your data.}

\item{random}{Named list of random effects. Each name should correspond to a variable in your data.}

\item{overdispersion}{List specifying the model for the overdispersion terms. This is limited to `gaussian_effect()` for now.}

\item{design}{List specifying the case control referent frames.}

\item{control_aghq}{List of parameters to be passed to the aghq::quad function.}
}
\value{
A list of lists of class "ccModel".
}
\description{
Constructor for ccModel (S3) objects.
}
\examples{
model <- ccModel(response = "count", time_index = "date",
                 fixed = list("hum_mean" = fixedEffect(gaussian_prior())),
                 random = list("o3_lag" = randomEffect(rw_effect(), pc_prec_prior()),
                               "temp_lag" = randomEffect(rw_effect(), pc_prec_prior())),
                 overdispersion = randomEffect(gaussian_effect(), pc_prec_prior(u = 10, alpha = .9)),
                 design = ccDesign(), aghq_input = aghqInput())
}
