% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{control_aghq}
\alias{control_aghq}
\title{Sets aghq parameters for fitting a model of model_type.}
\usage{
control_aghq(model_class, ...)
}
\arguments{
\item{model_class}{Either "cc" (case crossover) or "ts" (time series).}

\item{...}{See details.}
}
\value{
A list of parameters for use by aghq::aghq.
}
\description{
Sets aghq parameters for fitting a model of model_type.
}
\examples{
control_aghq(model_class = "cc", k = 3)
control_aghq(model_class = "ts")
}
