% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{createLaggedVariable}
\alias{createLaggedVariable}
\title{Creates a new_variable so that new_variables\link{t,} == variables\link{t-lag,}, where t is the time_index (e.g. a variable "date") of your data frame.}
\usage{
createLaggedVariable(
  df,
  variable,
  lags,
  remove_original = F,
  ext = paste0("_", lags, "lag")
)
}
\arguments{
\item{df}{A data.frame.}

\item{lags}{A integer vector giving the lags to use (must be of length 1 or same as variables).}

\item{remove_original}{A boolean indicating whether the variables should be removed from the data.frame. Default is FALSE.}

\item{ext}{A character vector indicating giving the names of the new_variables. Default is paste0("_", lags, "lag").}

\item{variables}{A character vector of variable names.}
}
\value{
An updated data.frame.
}
\description{
Creates a new_variable so that new_variables\link{t,} == variables\link{t-lag,}, where t is the time_index (e.g. a variable "date") of your data frame.
}
