% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitCCAGHQ}
\alias{fitCCAGHQ}
\title{Fits a case-crossover model}
\usage{
fitCCAGHQ(model, data, control_aghq = control_aghq(model_class = "cc"))
}
\arguments{
\item{df}{A data.frame.}

\item{variables}{A character vector of variable names.}

\item{lags}{A integer vector giving the lags to use (must be of length 1 or same as variables).}

\item{remove_original}{A boolean indicating whether the variables should be removed from the data.frame. Default is FALSE.}

\item{ext}{A character vector indicating giving the names of the new_variables. Default is paste0("_", lags, "lag").}
}
\value{
An updated data.frame.
}
\description{
Fits a case-crossover model
}
