% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fixedEffect}
\alias{fixedEffect}
\title{Specify a nonlinear random effect.}
\usage{
fixedEffect(variable, prior_dist = gaussian_prior(prec = 0.01))
}
\arguments{
\item{prior_dist}{A}
}
\description{
Specify a nonlinear random effect.
}
