% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Specify a gamma prior distribution.}
\usage{
gamma_prior(shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{prec}{TO DO.}
}
\value{
A list specifying a gamma prior distribution.
}
\description{
Specify a gamma prior distribution.
}
