% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{gaussian_effect}
\alias{gaussian_effect}
\title{Specify a Gaussian random effect.}
\usage{
gaussian_effect()
}
\value{
A list specifying a centered Gaussian distribution.
}
\description{
This is used to include overdispersion in the model. This is the only model implemented so far.
It does practically nothing except holding the character "gaussian".
}
