% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{gaussian_prior}
\alias{gaussian_prior}
\title{Specify a Gaussian prior distribution.}
\usage{
gaussian_prior(prec = 0.001)
}
\arguments{
\item{prec}{TO DO.}
}
\value{
A list specifying a Gaussian prior distribution.
}
\description{
Specify a Gaussian prior distribution.
}
