% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Extract results from a fitted ccModel.}
\usage{
getResults(
  fit,
  probs_pw = c(0.8, 0.9),
  probs_g = c(0.8, 0.9),
  M = 10000,
  values = NULL
)
}
\arguments{
\item{fit}{A fitted ccModel (obtained via `fitModel(model)`).}

\item{probs_pw}{Nominal levels for computing pointwise credible intervals for fixed and random effects (and overdispersion terms).}

\item{probs_g}{Nominal levels for computing global envelopes for fixed and random effects (and overdispersion terms).}

\item{M}{Number of samples used for inference.}

\item{values}{List of vectors containing the values at which to evaluate the splines (fixed effect), if included.}
}
\value{
A data.frame with values of the beta and gamma coefficients, as well as their credible intervals.
}
\description{
Extract results from a fitted ccModel.
}
