% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getResults_iwp}
\alias{getResults_iwp}
\title{Extract results from a fitted ccModel with IWP random effects.}
\usage{
getResults_iwp(fit, probs = 0.5, M = 10000)
}
\arguments{
\item{fit}{A fitted ccModel (obtained via \code{fitModel(model)}).}

\item{probs}{Quantiles of the posterior (cumulative) distribution of fixed and random effects (and overdispersion terms).}

\item{stepsizes}{Named vector specifying the (reported) stepsize for each random effect.}
}
\value{
A data.frame with values of the beta and gamma coefficients, as well as their credible intervals.
}
\description{
Extract results from a fitted ccModel with IWP random effects.
}
