% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults_z}
\alias{getResults_z}
\title{Extract (overdispersion) results from a fitted ccModel.}
\usage{
getResults_z(fit, probs_pw, probs_g, M, values)
}
\arguments{
\item{fit}{A fitted ccModel (obtained via \code{fitModel(model)}).}

\item{probs_pw}{Nominal levels for computing pointwise credible intervals for fixed and random effects (and overdispersion terms).}

\item{probs_g}{Nominal levels for computing global envelopes for fixed and random effects (and overdispersion terms).}

\item{M}{Number of samples used for inference.}
}
\value{
A data.frame with values of the beta and gamma coefficients, as well as their credible intervals.
}
\description{
Extract (overdispersion) results from a fitted ccModel.
}
