% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{iwp_effect}
\alias{iwp_effect}
\title{Specify an integrated Wiener process for random effect.}
\usage{
iwp_effect(
  order = 2,
  poly_degree = order - 1,
  ref_value = stats::median,
  knots
)
}
\arguments{
\item{order}{Order of the first non-zero derivative.}

\item{poly_degree}{Degree of the polynomial for additional fixed effects (by default `order`-1).
Note that the intercept is excluded.}

\item{ref_value}{Either a value or a function to be used for computing the reference value.}

\item{knots}{Number of (equally spaced) knots to use.}
}
\value{
A list specifying an integrated Wiener process model for a nonlinear random effect.
}
\description{
This is used to approximate Gaussian processes random effects.
By default, fixed effects (x - x_ref)^i, i=1,...,`order`-1
are added to the model.
}
\examples{
iwp_effect()
}
