% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{log_gamma_prior}
\alias{log_gamma_prior}
\title{Specify an log-gamma prior distribution.}
\usage{
log_gamma_prior(shape = 0.3, rate = 1e-06)
}
\arguments{
\item{alpha}{shape parameter.}

\item{beta}{rate parameter.}
}
\value{
A list specifying a log-gamma prior distribution.
}
\description{
Specify an log-gamma prior distribution.
}
\examples{
log_gamma_prior(alpha = .3, beta = 1e-6)
}
