% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{mgp_effect}
\alias{mgp_effect}
\title{Specify a montone Gaussian process for random effect.}
\usage{
mgp_effect(
  poly_degree = 1,
  ref_value = NULL,
  lambda = 2,
  c = 0.001,
  method = "FEM",
  knots = NULL,
  region = NULL,
  accuracy = 0.01,
  boundary = T
)
}
\arguments{
\item{poly_degree}{Degree of the polynomial for additional fixed effects (by default set to 1).
Note that the intercept is excluded.}

\item{ref_value}{Either a value or a function to be used for computing the reference value.}

\item{lambda}{Power of the Box-Cox type transformation.}

\item{c}{Small value added to the data for numerical reasons.}

\item{method}{Either `FEM` (finite elements, default) or `SS` (state space).}

\item{knots}{Number of (equally spaced) knots to use (for `method = 'FEM'`).}

\item{region}{Region of the covariate sample space to cover (for `method = 'FEM'`).}

\item{accuracy}{(Numeric. Default: `.001`).}

\item{boundary}{(Logical. Default: `T`).}
}
\value{
A list specifying a monotone Gaussian process process model for a nonlinear random effect.
}
\description{
By default, fixed effects (x - x_ref)^i, i=1,...,`poly_degree` (`poly_degree` = 1)
are added to the model.
}
\examples{
mgp_effect()
}
