% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{pc_prep_prior}
\alias{pc_prep_prior}
\title{Specify an exponential (Penalized Complexity) prior distribution.}
\usage{
pc_prep_prior(alpha = 0.5, u = 0.01)
}
\arguments{
\item{prec}{TO DO.}
}
\value{
A list specifying a pc prep prior distribution.
}
\description{
Specify an exponential (Penalized Complexity) prior distribution.
}
