% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{poly_effect}
\alias{poly_effect}
\title{Specify a polynomial fixed effect.}
\usage{
poly_effect(degree = 1, ref_value = median)
}
\arguments{
\item{poly_degree}{Degree of the polynomial.}
}
\value{
A list specifying a polynomial fixed effect.
}
\description{
Specify a polynomial fixed effect.
}
