% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_specifiers.R
\name{randomEffect}
\alias{randomEffect}
\title{Specify a nonlinear random effect.}
\usage{
randomEffect(model, theta_prior, beta_prior)
}
\arguments{
\item{model}{A model for the random effect: `rw_effect()` or `iwp_effect()`.}

\item{theta_prior}{Prior distribution for the parameters theta of the random effect model.}

\item{beta_prior}{Prior distribution for the fixed effects created as a by-product. Only the Gaussian prior is implemented.
The number of beta_prior parameters must match the number of coefficients created by the chosen model, i.e., `order-1`.
For example, a random walk of order 3 creates (by default) 2 fixed effects. In most cases, this argument should be left unspecified.}
}
\description{
Specify a nonlinear random effect.
}
\examples{
randomEffect(model = rw_effect(), theta_prior = pc_prec())
}
