% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{rw_effect}
\alias{rw_effect}
\title{Specify a Gaussian random walk distribution for random effect.}
\usage{
rw_effect(order = 3, poly_degree = order - 1, ref_value = median, binwidth = 1)
}
\arguments{
\item{order}{Order of the random walk.}

\item{poly_degree}{Degree of the polynomial interpolation at reference value; see details.}

\item{ref_value}{Either a value or a function to be used for computing the reference value.}

\item{binwidth}{Binwidth to be used for discretizing the continuous variable; see details.}
}
\value{
A list specifying a model for a nonlinear random effect.
}
\description{
Specify a Gaussian random walk distribution for random effect.
}
