% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_distributions.R
\name{rw_prior}
\alias{rw_prior}
\title{Specify a Gaussian random walk prior distribution.}
\usage{
rw_prior(order = 3, poly_degree = order - 1, ref_value = median, binwidth = 1)
}
\arguments{
\item{order}{Order of the random walk.}

\item{poly_degree}{Degree of the polynomial interpolation at reference value; see details.}

\item{ref_value}{Either a value or a function to be used for computing the reference value.}

\item{binwidth}{Binwidth to be used for discretizing the continuous variable; see details.}
}
\value{
A list specifying a model for the nonlinear random effect.
}
\description{
Specify a Gaussian random walk prior distribution.
}
