% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tsModel}
\alias{tsModel}
\title{Constructor for tsModel (S3) objects.}
\usage{
tsModel(response, time_index, fixed, random, overdispersion, control_aghq)
}
\arguments{
\item{.}{TO DO}
}
\value{
. TO DO
}
\description{
Constructor for tsModel (S3) objects.
}
\examples{
TO DO
}
