% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ccDesign}
\alias{ccDesign}
\title{Specify the case-crossover design to be used by fitModel.}
\usage{
ccDesign(...)
}
\arguments{
\item{...}{See details.}
}
\description{
Function for specifying the case control design.
}
\details{
A default design is returned by `ccDesign()`. This latter can be changed, e.g., with `ccDesign(lag = 2)`.
Arguments of interest are `scheme` ("unidirectional", "bidirectional", "time stratified"),
`n_control` (number of control days, for uni- and bidirectional schemes), `lag` (e.g., 7 for ensuring that
case and control fall on the same weekday) and `stratum_rule` (for time stratified schemes, the only rule
implemented is "month").
}
\examples{
ccDesign()
}
