% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ccModel}
\alias{ccModel}
\title{Constructor for ccModel (S3) objects.}
\usage{
ccModel(fixed, random, overdispersion, design, control_aghq)
}
\arguments{
\item{.}{TO DO}
}
\value{
. TO DO
}
\description{
Constructor for ccModel (S3) objects.
}
\examples{
TO DO
}
