% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineBetaGamma_iwp.R
\name{combineBetaGamma_iwp}
\alias{combineBetaGamma_iwp}
\title{Contruct o-spline from knots and boundary conditions.}
\usage{
combineBetaGamma_iwp(samps, model, grids = NULL, silent = FALSE)
}
\arguments{
\item{samps}{`$samps` output of `aghq::marginal_laplace_tmb()` or similar (see details).}

\item{model}{Object of class "ccModel" with iwp random effects (fitted with o-splines).}

\item{grids}{Named list of grids on which to evaluate the o-splines (see details).}
}
\value{
A modification of the samps object containing the o-splines evaluated at the given grid points.
}
\description{
Function used to evaluate random effects (o-splines, including fixed effects for boundary conditions) on a given grid.
}
\details{
`samps` is a matrix with samples as columns and rows that matches the fixed and random effects of the model.
If `grids` is NULL, then the grids for all random effects are constructed automatically.
If grids is a named list, then only the named random effects are considered;
if their corresponding grid is NULL, then it is constructed automatically.
}
