% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit a model}
\usage{
fitModel(model, data, silent = F, dll = NULL)
}
\arguments{
\item{model}{A model, either a ccModel or a tsModel object.}

\item{data}{The dataset on which the model is fitted. It should contain all the variables involved in the model.}

\item{silent}{A boolean indicating whether you wish to silence the \code{aghq::marginal_laplace_tmb()} function (default is \code{TRUE}).}
}
\value{
An updated ccModel or tsModel object that contains the fitted model (e.g., in ccModel$fit).
}
\description{
Function used for fitting either a ccModel (case crossover model) or a tsModel (time series model, yet to be implemented).
See the vignettes for examples of its usage.
}
