% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit a model}
\usage{
fitModel(model, data, silent = F, params_init = NULL)
}
\arguments{
\item{model}{A ccModel object.}

\item{data}{The dataset on which the model is fitted. It should contain all the variables involved in the model.}

\item{silent}{A boolean indicating whether you wish to silence the `aghq::marginal_laplace_tmb()` function (default is `TRUE`).}
}
\value{
An updated ccModel object that contains the fitted model (e.g., in ccModel$fit).
}
\description{
Function used for fitting a ccModel (case crossover model).
See the vignettes for examples of its usage.
}
