% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{randomEffect}
\alias{randomEffect}
\title{Specify a nonlinear random effect.}
\usage{
randomEffect(variable, model, prior)
}
\arguments{
\item{prior_dist}{A}
}
\description{
Specify a nonlinear random effect.
}
