% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{rw_effect}
\alias{rw_effect}
\title{Specify a Gaussian random walk distribution for random effect.}
\usage{
rw_effect(order = 3, poly_degree = order - 1, ref_value = median, binwidth = 1)
}
\arguments{
\item{order}{Order of the random walk. Restricted to 1, 2 and 3.}

\item{poly_degree}{Degree of the polynomial interpolation at reference value; see details.}

\item{ref_value}{Either a value or a function to be used for computing the reference value.}

\item{binwidth}{Binwidth to be used for discretizing the continuous variable; see details.}
}
\value{
A list specifying a random walk model for a nonlinear random effect.
}
\description{
This is used to approximate Gaussian processes random effects.
When a Gaussian random walk is specified, the range of observed values is discretized into
bins and the random effect is modeled as a random walk on these. For technical reasons, the
bin that contain the reference value is removed, as well as some of its neighbors, so that
the number of bins removed equals \code{order}. By default, fixed effects (x - x_ref)^i, i=1,...,\code{order}-1
are then added to the model.
}
\examples{
rw_effect()
}
